+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Displays : Tree is now a display
 - Files : Method from_file now sets Binary and StringFile filename attribute
+- Schemas : Order entry based on signature order
 
 ### Changed